Updates CommentIntenitySeeker.java and specifically its starttime method.
The start time method now searches the provided set of comments to find the most provitable starting point. A set is kept for all already looked up values to provide efficiency for larger sets with lots of duplicate starting times. The format is ugly because the GitHub editor does not auto indent and there is no tab on my currently used device.